+++ web/src/core/core.ts
@@ -3,6 +3,7 @@ import './utils.js';
 import { ExportType } from './export.js';
 import './sections.js';
 import './comments.js';
+import './items.js';
 import './fields.js';
 import './search.js';